Now possible to create early alloys in the kiln. Also added all possible alloys to the alloy manager.

+++ TFC_Shared/src/TFC/Core/Metal/AlloyMetal.java
@@ -5,9 +5,9 @@ import TFC.API.Metal;
 public class AlloyMetal 
 {
 	protected Metal metalType;
-	int metal;
+	float metal;
 	
-	public AlloyMetal(Metal e, int m)
+	public AlloyMetal(Metal e, float m)
 	{
 		metalType = e;
 		metal = m;
